// src/app/compartido/diseno/enums/uso-item-cir-rec.enum.ts
export enum UsoItemCircular {
    CIRPERFIL = 0, // item circular en el perfil
    CIRALBUM = 1, // Item circular en el album
    CIRCONTACTO = 2, // Item circular en el contacto
}

export enum UsoItemRectangular {
    RECPERFIL = 3, // Item rectangular en el perfil
    RECALBUMMINI = 4, // Item rectangular en el album - Para la miniatura
    RECALBUMPREVIEW = 5 // Item rectangular en el album - Para la preview
}